Add new files

+++ TEST/200646/lab6/lab6.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 #include <string>
 
-long double fun(double x)
+long double fun(long double x)
 {
     return x * x;
 }
